pyproject.toml
[tool.poetry]
name = "ai-summary-slack-bot"
version = "0.1.0"
description = ""
authors = ["kento.toriumi <kento.toriumi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
black = "^24.1.1"
newspaper3k = "^0.2.8"
google-generativeai = "^0.3.2"
slack-sdk = "^3.26.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
